cast_paligemma/requirements.txt
# Core ML libraries
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
pillow>=9.0.0
opencv-python>=4.8.0

# Visualization and evaluation
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0

# Configuration and logging
pyyaml>=6.0
wandb>=0.15.0

# Additional utilities
scikit-learn>=1.3.0
scipy>=1.10.0